// server/prisma/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const JobType = {
  FULL_TIME: 'FULL_TIME',
  PART_TIME: 'PART_TIME',
  REMOTE: 'REMOTE',
  HYBRID: 'HYBRID',
  CONTRACT: 'CONTRACT',
  INTERNSHIP: 'INTERNSHIP',
  ONSITE: 'ONSITE'
} as const

export type JobType = (typeof JobType)[keyof typeof JobType]


export const Role = {
  USER: 'USER',
  ADMIN: 'ADMIN',
  EMPLOYER: 'EMPLOYER'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const AuthProvider = {
  CREDENTIALS: 'CREDENTIALS',
  GOOGLE: 'GOOGLE',
  FACEBOOK: 'FACEBOOK'
} as const

export type AuthProvider = (typeof AuthProvider)[keyof typeof AuthProvider]


export const AppStatus = {
  PENDING: 'PENDING',
  REVIEWED: 'REVIEWED',
  ACCEPTED: 'ACCEPTED',
  REJECTED: 'REJECTED'
} as const

export type AppStatus = (typeof AppStatus)[keyof typeof AppStatus]